requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
colorama==0.4.5
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
Faker==12.0.1
Markdown==3.4.1
mixer==7.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
